% File src/library/datasets/man/austres.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1994-9 W. N. Venables and B. D. Ripley
% Distributed under GPL 2 or later

\name{austres}
\docType{data}
\alias{austres}
\title{
  Quarterly Time Series of the Number of Australian Residents
}
\description{
  Numbers (in thousands) of Australian residents measured quarterly from
  March 1971 to March 1994.  The object is of class \code{"ts"}.
}
\usage{austres}
\source{
  P. J. Brockwell and R. A. Davis (1996)
  \emph{Introduction to Time Series and Forecasting.}
  Springer
}
\keyword{datasets}
