% File src/library/datasets/man/BJsales.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2025 R Core Team
% Distributed under GPL 2 or later

\name{BJsales}
\docType{data}
\alias{BJsales}
\alias{BJsales.lead}
\title{Sales Data with Leading Indicator}
\description{
  The sales time series \code{BJsales} and leading indicator
  \code{BJsales.lead} each contain 150 observations.
  The objects are of class \code{"ts"}.
}
\usage{
BJsales
BJsales.lead
}
\source{
  The data are given in \bibcitet{|R:Box+Jenkins:1976|p.\sspace{}537}.
  Obtained from the Time Series Data Library at
  \url{https://robjhyndman.com/TSDL/}
}
\references{
  \bibinfo{R:Brockwell+Davis:1991}{note}{Page 414}
  \bibshow{*, R:Brockwell+Davis:1991}
}
\keyword{datasets}
