% File src/library/datasets/man/datasets-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{datasets-package}
\alias{datasets-package}
\alias{datasets}
\docType{package}
\title{
The R Datasets Package
}
\description{
Base R datasets
}
\details{This package contains a variety of datasets.  For a complete
list, use \code{library(help = "datasets")}.
}
\author{
R Core Team and contributors worldwide

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
