\name{datasets.package}
\alias{datasets.package}
\alias{datasets}
\docType{package}
\title{
The R Datasets Package
}
\description{
Base R datasets
}
\details{This package contains a variety of datasets.  For a complete
list, use \code{library(help="datasets")}.
}
\author{
R Development Core Team and contributors worldwide

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
