\name{iris}
\alias{iris}
\title{Fisher's Iris Data}
\description{
A classic 150-row classification dataset of iris flower measurements. It is one
of the most widely used teaching datasets in statistics and machine learning.
}
\format{
A CSV file with 150 rows and 5 columns:
\describe{
  \item{sepal_length_cm}{Sepal length in centimeters.}
  \item{sepal_width_cm}{Sepal width in centimeters.}
  \item{petal_length_cm}{Petal length in centimeters.}
  \item{petal_width_cm}{Petal width in centimeters.}
  \item{species}{Species label.}
}
}
\details{
Local file: \file{datasets/data/iris.csv}. The repository stores a normalized
CSV copy with a header row added to the UCI source file.
}
\source{
UCI Machine Learning Repository dataset page:
\url{https://archive.ics.uci.edu/dataset/53/iris}.
Original download:
\url{https://archive.ics.uci.edu/static/public/53/iris.zip}.
License: CC BY 4.0.
}
\references{
Fisher, R. A. (1936). Iris [Dataset]. UCI Machine Learning Repository.
\url{https://doi.org/10.24432/C56C76}.
}
\keyword{datasets}
