% File src/library/datasets/man/lynx.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2025 R Core Team
% Distributed under GPL 2 or later

\name{lynx}
\docType{data}
\alias{lynx}
\title{Annual Canadian Lynx trappings 1821--1934}
\description{
  Annual numbers of lynx trappings for 1821--1934 in Canada.
  Taken from \bibcitet{R:Brockwell+Davis:1991}, this appears to be the
  series considered by \bibcitet{R:Campbell+Walker:1977}.
}
\usage{lynx}
\source{
  \bibinfo{R:Brockwell+Davis:1991}{note}{Series G (page 557)}
  \bibshow{R:Brockwell+Davis:1991}
}
\references{
  \bibshow{R:Becker+Chambers+Wilks:1988, R:Campbell+Walker:1977}
}
\keyword{datasets}
