% File src/library/datasets/man/sunspot.year.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\newcommand{\sspace}{\ifelse{latex}{\out{~}}{ }}

\name{sunspot.year}
\docType{data}
\alias{sunspot.year}
\title{
  Yearly Sunspot Data, 1700--1988
}
\description{
  Yearly numbers of sunspots.
}
\usage{
sunspot.year
}
\format{
  The univariate time series \code{sunspot.year} contains 289
  observations, and is of class \code{"ts"}.
}
\source{
  H. Tong (1996)
  \emph{Non-Linear Time Series}. Clarendon Press, Oxford, p.\sspace{}471.
}
\keyword{datasets}
