% File src/library/datasets/man/UKgas.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2026 R Core Team
% Distributed under GPL 2 or later

\name{UKgas}
\docType{data}
\alias{UKgas}
\title{UK Quarterly Gas Consumption}
\usage{UKgas}
\description{
  Quarterly UK gas consumption from 1960Q1 to 1986Q4, in millions of therms.
}
\format{
  A quarterly time series of length 108.
}
\source{
  \bibshow{R:Durbin+Koopman:2001}\cr
  Data was available from the book website, now archived at
  \url{https://web.archive.org/web/20020304222057/http://www.ssfpack.com/dkbook/}.
}
\examples{
## maybe str(UKgas) ; plot(UKgas) ...
}
\keyword{datasets}
