\name{wine}
\alias{wine}
\title{Wine Recognition Data}
\description{
A 178-row multiclass dataset of chemical measurements from wines grown in the
same region of Italy but derived from three cultivars.
}
\format{
A CSV file with 178 rows and 14 columns:
\describe{
  \item{class}{Wine class identifier from 1 to 3.}
  \item{alcohol}{Alcohol content.}
  \item{malic_acid}{Malic acid concentration.}
  \item{ash}{Ash content.}
  \item{alcalinity_of_ash}{Alcalinity of ash.}
  \item{magnesium}{Magnesium content.}
  \item{total_phenols}{Total phenols.}
  \item{flavanoids}{Flavanoids.}
  \item{nonflavanoid_phenols}{Nonflavanoid phenols.}
  \item{proanthocyanins}{Proanthocyanins.}
  \item{color_intensity}{Color intensity.}
  \item{hue}{Hue.}
  \item{od280_od315}{OD280/OD315 of diluted wines.}
  \item{proline}{Proline concentration.}
}
}
\details{
Local file: \file{datasets/data/wine.csv}. The repository stores a normalized
CSV copy with a header row added to the UCI source file.
}
\source{
UCI Machine Learning Repository dataset page:
\url{https://archive.ics.uci.edu/dataset/109/wine}.
Original download:
\url{https://archive.ics.uci.edu/static/public/109/wine.zip}.
License: CC BY 4.0.
}
\references{
Aeberhard, S. and Forina, M. (1992). Wine [Dataset]. UCI Machine Learning
Repository. \url{https://doi.org/10.24432/C5PC7J}.
}
\keyword{datasets}
