% File src/library/datasets/man/BJsales.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{BJsales}
\docType{data}
\alias{BJsales}
\alias{BJsales.lead}
\title{Sales Data with Leading Indicator}
\description{
  The sales time series \code{BJsales} and leading indicator
  \code{BJsales.lead} each contain 150 observations.
  The objects are of class \code{"ts"}.
}
\usage{
BJsales
BJsales.lead
}
\source{
  The data are given in Box & Jenkins (1976).
  Obtained from the Time Series Data Library at
  \url{http://www-personal.buseco.monash.edu.au/~hyndman/TSDL/}
}
\references{
  G. E. P. Box and G. M. Jenkins (1976):
  \emph{Time Series Analysis, Forecasting and Control},
  Holden-Day, San Francisco, p.\sspace{}537.

  P. J. Brockwell and R. A. Davis (1991):
  \emph{Time Series: Theory and Methods},
  Second edition, Springer Verlag, NY, pp.\sspace{}414.
}
\keyword{datasets}
