% File src/library/datasets/man/eurodist.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{eurodist}
\docType{data}
\alias{eurodist}
\title{Distances Between European Cities}
\description{
  The data give the road distances (in km) between 21 cities in Europe.
  The data are taken from a table in \emph{The Cambridge Encyclopaedia}.
}
\usage{eurodist}
\format{
  A \code{dist} object based on 21 objects.
  (You must have the \pkg{stats} package loaded to have the methods for this
  kind of object available).
}
\source{
  Crystal, D. Ed. (1990)
  \emph{The Cambridge Encyclopaedia}.
  Cambridge: Cambridge University Press,
}
\keyword{datasets}
