% File src/library/datasets/man/LakeHuron.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{LakeHuron}
\docType{data}
\alias{LakeHuron}
\title{Level of Lake Huron 1875--1972}
\description{
  Annual measurements of the level, in feet, of Lake Huron 1875--1972.
}
\usage{LakeHuron}
\format{
  A time series of length 98.
}
\source{
\bibinfo{R:Brockwell+Davis:1991}{note}{Page 555}
\bibinfo{R:Brockwell+Davis:1996}{note}{Sections 5.1 and 7.6}
\bibshow{R:Brockwell+Davis:1991,R:Brockwell+Davis:1996}
}
\keyword{datasets}
