% File src/library/datasets/man/lh.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1994-9 W. N. Venables and B. D. Ripley
% Distributed under GPL 2 or later

\name{lh}
\docType{data}
\alias{lh}
\title{
  Luteinizing Hormone in Blood Samples
}
\description{
  A regular time series giving the luteinizing hormone in blood
  samples at 10 \abbr{min}s intervals from a human female, 48 samples.
}
\usage{lh}
\source{
  \bibinfo{R:Diggle:1991}{note}{Table A.1, series 3}
  \bibshow{R:Diggle:1991}
}
\keyword{datasets}
