% File src/library/datasets/man/treering.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{treering}
\docType{data}
\alias{treering}
\title{
  Yearly Treering Data, -6000--1979
}
\description{
  Contains normalized tree-ring widths in dimensionless units.
}
\usage{
treering
}
\format{
  A univariate time series with 7981 observations. The object is of
  class \code{"ts"}.

  Each tree ring corresponds to one year.
}
\details{
  The data were recorded by Donald A. Graybill, 1980, from
  Gt Basin Bristlecone Pine 2805M, 3726-11810 in Methuselah Walk, California.
}
\source{
  Time Series Data Library:
  \url{http://www-personal.buseco.monash.edu.au/~hyndman/TSDL/},
  series \file{CA535.DAT}
}
\references{
  For background on Bristlecone pines and Methuselah Walk, see
  \url{http://www.sonic.net/bristlecone/}; for some photos see
  \url{http://www.ltrr.arizona.edu/~hallman/sitephotos/meth.html}
}
\keyword{datasets}
