% File src/library/datasets/man/UKgas.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{UKgas}
\docType{data}
\alias{UKgas}
\title{UK Quarterly Gas Consumption}
\usage{UKgas}
\description{
  Quarterly UK gas consumption from 1960Q1 to 1986Q4, in millions of therms.
}
\format{
  A quarterly time series of length 108.
}
\source{
  \bibshow{R:Durbin+Koopman:2001}
}
\examples{
## maybe str(UKgas) ; plot(UKgas) ...
}
\keyword{datasets}
